refactor: snackbar testable 구조로 변경
